Create Configuration component and move logic into there.
