+++ src/features/import/components/Importer/Configure/index.tsx
@@ -1,12 +1,9 @@
-import { CompareArrows } from '@mui/icons-material';
-import { Box, useTheme } from '@mui/material';
+import { Box } from '@mui/material';
 import { FC, useState } from 'react';
 
+import Configuration from './Configuration';
 import { ExperimentalFieldTypes } from './Mapping/MappingRow';
 import Mapping from './Mapping';
-import messageIds from 'features/import/l10n/messageIds';
-import { useMessages } from 'core/i18n';
-import ZUIEmptyState from 'zui/ZUIEmptyState';
 import SheetSettings, { ExperimentSheet } from './SheetSettings';
 
 export interface ConfiguringData {
@@ -19,9 +16,6 @@ interface ConfigureProps {
 }
 
 const Configure: FC<ConfigureProps> = ({ sheets }) => {
-  const messages = useMessages(messageIds);
-  const theme = useTheme();
-
   //settings
   const [firstRowIsHeaders, setFirstRowIsHeaders] = useState(true);
   const [selectedSheetId, setSelectedSheetId] = useState(sheets[0].id);
@@ -73,21 +67,7 @@ const Configure: FC<ConfigureProps> = ({ sheets }) => {
           />
         </Box>
         <Box width="50%">
-          {currentlyConfiguring && <>Mapping</>}
-          {!currentlyConfiguring && (
-            <Box
-              alignItems="center"
-              bgcolor={theme.palette.transparentGrey.light}
-              display="flex"
-              height="100%"
-              justifyContent="center"
-            >
-              <ZUIEmptyState
-                message={messages.configuration.mapping.emptyStateMessage()}
-                renderIcon={(props) => <CompareArrows {...props} />}
-              />
-            </Box>
-          )}
+          <Configuration currentlyConfiguring={currentlyConfiguring} />
         </Box>
       </Box>
       <Box padding={4}>Preview</Box>
